+++ kernel/kernel/Process.cpp
@@ -1997,7 +1997,7 @@ namespace Kernel
 		return TRY(m_open_file_descriptors.dup2(fildes, fildes2));
 	}
 
-	BAN::ErrorOr<long> Process::sys_fcntl(int fildes, int cmd, int extra)
+	BAN::ErrorOr<long> Process::sys_fcntl(int fildes, int cmd, uintptr_t extra)
 	{
 		return TRY(m_open_file_descriptors.fcntl(fildes, cmd, extra));
 	}
